shaders: Update common_iostructs.

+++ res/gamedata/shaders/gl/common_iostructs.h
@@ -277,9 +277,9 @@ struct         v_static
 	float4	Nh		; // NORMAL;	// (nx,ny,nz,hemi occlusion)
 	float4	T		; // TANGENT;	// tangent
 	float4	B		; // BINORMAL;	// binormal
-	int2	tc		; // TEXCOORD0;	// (u,v)
+	float2	tc		; // TEXCOORD0;	// (u,v)
 #ifdef	USE_LM_HEMI
-	int2	lmh		; // TEXCOORD1;	// (lmu,lmv)
+	float2	lmh		; // TEXCOORD1;	// (lmu,lmv)
 #endif
 //	float4	color	; // COLOR0;	// (r,g,b,dir-occlusion)	//	Swizzle before use!!!
 	float4	P		; // POSITION;	// (float,float,float,1)
@@ -290,9 +290,9 @@ struct	v_static_color
 	float4	Nh		; // NORMAL;	// (nx,ny,nz,hemi occlusion)
 	float4	T		; // TANGENT;	// tangent
 	float4	B		; // BINORMAL;	// binormal
-	int2	tc		; // TEXCOORD0;	// (u,v)
+	float2	tc		; // TEXCOORD0;	// (u,v)
 #ifdef	USE_LM_HEMI
-	int2	lmh		; // TEXCOORD1;	// (lmu,lmv)
+	float2	lmh		; // TEXCOORD1;	// (lmu,lmv)
 #endif
 	float4	color	; // COLOR0;	// (r,g,b,dir-occlusion)	//	Swizzle before use!!!
 	float4	P		; // POSITION;	// (float,float,float,1)
